LiberalPython/LiberalPython/requirements.txt
click==7.1.2
Flask==1.1.4
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.9.1
MarkupSafe==2.0.1
pip==20.1.1
python-docx==0.8.11
setuptools==47.1.0
Werkzeug==1.0.1
simpletransformers
scikit-learn